backend/requirements.txt
Django>=4.2
mysqlclient
djangorestframework
python-dotenv
drf-spectacular
django-cors-headers